+++ src/main/php/ConsoleApp.php
@@ -117,6 +117,20 @@ abstract class ConsoleApp extends App
     }
 
     /**
+     * creates list of bindings from given class
+     *
+     * @param   string  $className    full qualified class name of class to create an instance of
+     * @param   string  $projectPath  path to project
+     * @return  \stubbles\ioc\module\BindingModule[]
+     */
+    protected static function getBindingsForApp($className, $projectPath)
+    {
+        $bindings   = parent::getBindingsForApp($className, $projectPath);
+        $bindings[] = new ConsoleBindingModule();
+        return $bindings;
+    }
+
+    /**
      * creates argument binding module
      *
      * @api
@@ -132,6 +146,7 @@ abstract class ConsoleApp extends App
      *
      * @api
      * @return  \stubbles\console\ioc\ConsoleBindingModule
+     * @deprecated  since 4.0.0, console binding module will be added by default, will be removed with 5.0.0
      */
     protected static function createConsoleBindingModule()
     {