+++ src/test/php/ConsoleAppTest.php
@@ -9,9 +9,6 @@
  */
 namespace stubbles\console;
 use stubbles\streams\memory\MemoryOutputStream;
-require_once __DIR__ . '/ConsoleAppUsingBindingModule.php';
-require_once __DIR__ . '/SelfBoundConsoleApp.php';
-require_once __DIR__ . '/TestConsoleApp.php';
 use org\stubbles\console\test\ConsoleAppUsingBindingModule;
 use org\stubbles\console\test\SelfBoundConsoleApp;
 use org\stubbles\console\test\TestConsoleApp;