+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * change of annotation value names in request broker for user input classes: `name` must now be `paramName`, group must now be `paramGroup`
   * changed all thrown stubbles/core exceptions to those recommended with stubbles/core 5.0.0
   * removed `stubbles\console\Executor::getOutputStream()`, use `stubbles\console\Executor::out()` instead, was deprecates since 3.0.0
+  * deprecated `stubbles\console\ConsoleApp::createConsoleBindingModule()`, console binding module will be added by default, will be removed with 5.0.0
 
 
 ### Other changes